🏷️ refactor(whiteboard): change `id` type to `string`

+++ services/whiteboard-service/internal/converter/whiteboard/to_proto.go
@@ -7,7 +7,7 @@ import (
 
 func ToProtoWhiteboard(entity *entity.Whiteboard) *whiteboardv1.Whiteboard {
 	return &whiteboardv1.Whiteboard{
-		Id:          int64(entity.ID),
+		Id:          entity.ID,
 		Title:       entity.Title,
 		Description: entity.Description,
 	}
